Create apply_geo.csh
The following automated csh script will first apply the translation and then the rotation for all particles in a docfile. This script should be run from the same directory as from where the mltomo program was launched (so that the relative filenames in the docfile point to the correct files).If one saves this script as apply_geo.csh and makes it executable, the syntaxis for running it would be:
./apply_geo.csh MLtomo/run1/mltomo_it000010.doc app
This would create a copy of each aligned sub-tomogram with an additional extension .app.