Create prep_conic_2.4
The bellow scripts assumes that you start the conical tilt experiment with Xmipp and with the default options (you store micrographs in the “Preprocessing” directory and extracted images in the “Images” directory). When you finished picking in Xmipp, extract and process both the tilted and untilted images with “Preprocess Particles” still in Xmipp. Then run the ”prep_conic.b” script. This will create in the Xmipp home folder an untilted (untilt_stack.img) and a tilted stack (tilt_stack.img).